Improved input binding time (#49)
Improved the binding time of input, which should reduce the amount of code GHC has to deal with somewhat. This also allows for the construction of some improved operations for interacting with input fetching -- this will be useful on #46. 

There is further work that can be done here, since a static/dynamic `Rep` for input can be made and used to further optimise the operations: this will come soon too.

As a by-product, this PR fixes #45, though it does not _remove_ `Text16`, it changes it to just act like `Text`, making it entirely benign: `Text16` itself will be deprecated.